Make folder structure flat
