--- api/consts/error.js
@@ -1,6 +1,0 @@
-const error = {
-    getUsers: 'the users could not be retrieved',
-    getUserById: 'the user could not be retrieved'
-}
-
-module.exports = error;
