--- api/consts/url.js
@@ -1,6 +1,0 @@
-const url = {
-    users: '/api/users',
-    usersById: '/api/users/:id'
-}
-
-module.exports = url;
