--- dummyData.js/users.js
@@ -1,16 +1,0 @@
-const users  = [
-    {
-        id: 1,
-        name: 'Gabe',
-    }, 
-    {
-        id: 2,
-        name: 'Luke',
-    },
-    {
-        id: 3,
-        name: 'Tom',
-    },
-]
-
-module.exports = users;
