+++ utils/error.js
@@ -1,0 +1,6 @@
+const error = {
+  getUsers: 'the users could not be retrieved',
+  getUserById: 'the user could not be retrieved'
+};
+
+module.exports = error;
