+++ utils/url.js
@@ -1,0 +1,6 @@
+const url = {
+  users: '/api/users',
+  usersById: '/api/users/:id'
+};
+
+module.exports = url;
